+++ src/plugins/api-ai.plugin.ts
@@ -1,11 +1,13 @@
 import { EventEmitter } from 'events';
-import { removeWordsFromMessage, checkWordsInMessage } from '../lib/utils';
-import { IBotPlugin, IBotMessage, IBot } from '../lib/interfaces';
+
+import { BaseBotPlugin } from '../lib/base.bot-plugin';
+import { IBot, IBotMessage } from '../lib/interfaces';
+import { removeWordsFromMessage } from '../lib/utils';
 
 
 const apiai = require('apiai');
 
-export class ApiAiBotPlugin implements IBotPlugin {
+export class ApiAiBotPlugin extends BaseBotPlugin {
     public name = 'api-ai';
     public description = 'Simple usage https://api.ai service with default agent';
     public whatCanIdo = {
@@ -13,32 +15,27 @@ export class ApiAiBotPlugin implements IBotPlugin {
         'ru': 'Умею отвечать на простые вопросы, пример: `Как дела?`'
     };
     protected wordsForSpy: string[];
-    protected ai: any;
+    protected ai: any = {
+        'en': null,
+        'ru': null
+    }
+    protected ai_ru: any;
     constructor(
+        protected botLocale: string,
         protected botNameAliases: string[],
         protected apiaiClientAccessToken: string
     ) {
+        super(botLocale, botNameAliases);
         this.wordsForSpy = botNameAliases;
-        this.ai = apiai(apiaiClientAccessToken);
+        this.ai['en'] = apiai(apiaiClientAccessToken, { language: 'en' });
+        this.ai['ru'] = apiai(apiaiClientAccessToken, { language: 'ru' });
     }
-    public check(bot: IBot, msg: IBotMessage): boolean {
-        return msg.text &&
-            (
-                checkWordsInMessage(msg.text, this.wordsForSpy) ||
-                msg.chat.type === 'private'
-            );
-    }
-    public answerWhatCanIdo(bot: IBot, msg: IBotMessage): string {
-        if (msg.from && msg.from.language_code && msg.from.language_code.toLowerCase().indexOf('ru') !== -1) {
-            return this.whatCanIdo['ru'];
-        }
-        return this.whatCanIdo['en'];
-    }
-    protected askAi(message: string, sessionId: string): EventEmitter {
+    protected askAi(message: string, sessionId: string, locale: string): EventEmitter {
         const event = new EventEmitter();
         try {
-            const request = this.ai.textRequest(message.substring(0, 255), {
-                sessionId: sessionId
+            const request = this.ai[locale].textRequest(message.substring(0, 255), {
+                sessionId: sessionId,
+                lang: locale
             });
             request.on('response', function (response: any) {
                 event.emit('message', response.result.fulfillment.speech);
@@ -52,12 +49,12 @@ export class ApiAiBotPlugin implements IBotPlugin {
     public process(bot: IBot, msg: IBotMessage): EventEmitter {
         const event = new EventEmitter();
         if (msg.text) {
-            this.askAi(msg.text, msg.chat.id).on('message', (answer: string) => {
+            this.askAi(msg.text, msg.chat.id, this.getLocaleCode(msg)).on('message', (answer: string) => {
                 if (answer) {
                     event.emit('message', answer);
                 } else {
                     msg.text = removeWordsFromMessage(msg.text, this.wordsForSpy);
-                    this.askAi(msg.text, msg.chat.id).on('message', (answerTwo: string) => {
+                    this.askAi(msg.text, msg.chat.id, this.getLocaleCode(msg)).on('message', (answerTwo: string) => {
                         event.emit('message', answerTwo);
                     })
                 }