+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/EndyKaufman/kaufman-bot/compare/1.0.1...1.2.0) (2017-10-23)
+
+
+### Bug Fixes
+
+* Small error on debug microsoft bot ([bc06217](https://github.com/EndyKaufman/kaufman-bot/commit/bc06217))
+
+
+### Features
+
+* **api-ai:** Update for multi language agents ([0f1d0ee](https://github.com/EndyKaufman/kaufman-bot/commit/0f1d0ee))
+
+
+
 <a name="1.1.0"></a>
 # 1.1.0 (2017-08-19)